Allow MIDI IN callbacks to know the source device
The read() callbacks do not have a field for user
data, so there is no way to pass a pointer to the
device object to a read() callback. This commit
modifies the readAll() function to record the
device and cable number right before the call
to read(). If the call to read results in a
read callback being called, the callback can
call the new function getCurrentReadDevAndCable()
to identify which device and virtual cable
received data. This information is enough to
route data to another interface, or to grab
a pointer to the associated device or interface object.